MNT Initial update to smd config templates. Attempt at more robust psana2 env var calculation.
